+++ packages/composites/apy-finance/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @chainlink/apy-finance-adapter
 
+## 2.0.26
+
+### Patch Changes
+
+- Updated dependencies [5916b71df]
+  - @chainlink/token-allocation-adapter@1.10.2
+
 ## 2.0.25
 
 ### Patch Changes