+++ packages/composites/curve-3pool/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @chainlink/curve-3pool-adapter
 
+## 3.0.26
+
+### Patch Changes
+
+- Updated dependencies [5916b71df]
+  - @chainlink/token-allocation-adapter@1.10.2
+
 ## 3.0.25
 
 ### Patch Changes