+++ packages/sources/renvm-address-set/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @chainlink/renvm-address-set-adapter
 
+## 1.5.36
+
+### Patch Changes
+
+- Updated dependencies [5916b71df]
+  - @chainlink/proof-of-reserves-adapter@1.16.7
+
 ## 1.5.35
 
 ### Patch Changes